Make OSIAM version configurable
Add property for version with default value of 3, so that the upcoming
mainline version is supported by default. Check only the first character
of the version for '2' to provide maximum flexibility. Use OSIAM 3 for
any other version because it's the default. Inline the method
`PluginImpl#getProperty`, because there's no need to wrap
`System#getProperty`.